scouting/requirements.txt
asgiref==3.8.1
Django==5.1.2
sqlparse==0.5.1
django-environ
gunicorn
psycopg2-binary
requests
deepdiff
python-dateutil